Relocate negate pipe to utilities
