+++ src/app/utilities/negate.pipe.spec.ts
