+++ src/app/utilities/negate.pipe.ts
